requirements.txt
aniso8601==8.1.0
blinker==1.4
click==7.1.2
colorzero==1.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-DebugToolbar==0.11.0
Flask-JWT-Extended==3.25.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
future==0.18.2
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
picamera==1.13
pkg-resources==0.0.0
py-qmc5883l==0.1.1
PyJWT==1.7.1
pyserial==3.5
pytz==2020.5
PyYAML==5.4.1
quick2wire-api==0.0.0.2
RPi.GPIO==0.7.0
six==1.15.0
smbus==1.1.post2
SQLAlchemy==1.3.22
Werkzeug==1.0.1
